refactor: move the decoupling components out of Signal into their domains
