+++ dpsim-models/include/dpsim-models/EMT/EMT_Ph1_DecouplingIdealTransformer.h
@@ -6,36 +6,36 @@
 #include <vector>
 
 #include "dpsim-models/Definitions.h"
-#include "dpsim-models/SP/SP_Ph1_ControlledVoltageSource.h"
-#include "dpsim-models/SimNode.h"
-#include "dpsim-models/TopologicalNode.h"
+#include "dpsim-models/EMT/EMT_Ph1_VoltageSource.h"
 #include <dpsim-models/CompositePowerComp.h>
-#include <dpsim-models/SP/SP_Ph1_ControlledCurrentSource.h>
-#include <dpsim-models/SP/SP_Ph1_Resistor.h>
+#include <dpsim-models/EMT/EMT_Ph1_CurrentSource.h>
+#include <dpsim-models/EMT/EMT_Ph1_Resistor.h>
 
 namespace CPS {
-namespace Signal {
-class DecouplingIdealTransformer_SP_Ph1
-    : public CompositePowerComp<Complex>,
-      public SharedFactory<DecouplingIdealTransformer_SP_Ph1> {
+namespace EMT {
+namespace Ph1 {
+
+class DecouplingIdealTransformer
+    : public CompositePowerComp<Real>,
+      public SharedFactory<DecouplingIdealTransformer> {
 protected:
   Real mDelay;
   Real mInternalSeriesResistance = 1e-6;
   Real mInternalParallelResistance = 1e6;
 
-  std::shared_ptr<SP::Ph1::Resistor> mRes1, mRes2;
-  std::shared_ptr<SP::Ph1::ControlledCurrentSource> mCurrentSrc;
-  std::shared_ptr<SP::Ph1::ControlledVoltageSource> mVoltageSrc;
+  std::shared_ptr<EMT::Ph1::Resistor> mRes1, mRes2;
+  std::shared_ptr<EMT::Ph1::CurrentSource> mCurrentSrc;
+  std::shared_ptr<EMT::Ph1::VoltageSource> mVoltageSrc;
   Attribute<Complex>::Ptr mSrcCurrent, mSrcVoltage;
 
   // Ringbuffers for the values of previous timesteps
   // TODO make these matrix attributes
-  std::vector<Complex> mCur1, mVol2;
+  std::vector<Real> mCur1, mVol2;
 
   // Copy of the most recent elements of the ring buffers
   // They are used to perform extrapolation
-  std::vector<Complex> mCur1Extrap, mVol2Extrap;
-  Complex mCurrent1Extrap0;
+  std::vector<Real> mCur1Extrap, mVol2Extrap;
+  Real mCurrent1Extrap0;
 
   UInt mBufIdx = 0;
   UInt mMacroBufIdx = 0;
@@ -43,33 +43,33 @@ protected:
   Real mAlpha;
   CouplingMethod mCouplingMethod;
   UInt mExtrapolationDegree = 0;
-  Matrix mVoltageSrcIntfCurr;
+  Eigen::MatrixXd mVoltageSrcIntfCurr;
 
   // Get an approximate value of the signal in between steps when the delay is not an integer multiple of the step size
-  Complex interpolate(std::vector<Complex> &data);
+  Real interpolate(std::vector<Real> &data);
 
   // Estimates the value of the input signal in the next step
-  Complex extrapolate(std::vector<Complex> &data);
+  Real extrapolate(std::vector<Real> &data);
 
 public:
-  typedef std::shared_ptr<DecouplingIdealTransformer_SP_Ph1> Ptr;
+  typedef std::shared_ptr<DecouplingIdealTransformer> Ptr;
 
-  const Attribute<Complex>::Ptr mSourceVoltageIntfVoltage;
-  const Attribute<Complex>::Ptr mSourceVoltageIntfCurrent;
-  const Attribute<Complex>::Ptr mSrcCurrentRef;
-  const Attribute<Complex>::Ptr mSrcVoltageRef;
+  const Attribute<Real>::Ptr mSourceVoltageIntfVoltage;
+  const Attribute<Real>::Ptr mSourceVoltageIntfCurrent;
+  const Attribute<Real>::Ptr mSrcCurrentRef;
+  const Attribute<Real>::Ptr mSrcVoltageRef;
 
   ///FIXME: workaround for dependency analysis as long as the states aren't attributes
   const Attribute<Matrix>::Ptr mStates;
 
-  DecouplingIdealTransformer_SP_Ph1(
-      String uid, String name, Logger::Level logLevel = Logger::Level::info);
-  DecouplingIdealTransformer_SP_Ph1(
-      String name, Logger::Level logLevel = Logger::Level::info)
-      : DecouplingIdealTransformer_SP_Ph1(name, name, logLevel) {}
+  DecouplingIdealTransformer(String uid, String name,
+                             Logger::Level logLevel = Logger::Level::info);
+  DecouplingIdealTransformer(String name,
+                             Logger::Level logLevel = Logger::Level::info)
+      : DecouplingIdealTransformer(name, name, logLevel) {}
 
   void setParameters(Real delay, Matrix voltageSrcIntfCurr,
-                     Complex current1Extrap0,
+                     Real current1Extrap0,
                      CouplingMethod method = CouplingMethod::DELAY);
   void step(Real time, Int timeStepCount);
   void postStep();
@@ -96,5 +96,6 @@ public:
   void mnaCompUpdateVoltage(const Matrix &leftVector) override;
   void mnaCompUpdateCurrent(const Matrix &leftVector) override;
 };
-} // namespace Signal
+} // namespace Ph1
+} // namespace EMT
 } // namespace CPS
