+++ dpsim-models/include/dpsim-models/EMT/EMT_Ph3_DecouplingIdealTransformer.h
@@ -6,35 +6,39 @@
 #include <vector>
 
 #include "dpsim-models/Definitions.h"
-#include "dpsim-models/EMT/EMT_Ph1_VoltageSource.h"
+#include "dpsim-models/EMT/EMT_Ph3_ControlledVoltageSource.h"
+#include "dpsim-models/MathUtils.h"
 #include <dpsim-models/CompositePowerComp.h>
-#include <dpsim-models/EMT/EMT_Ph1_CurrentSource.h>
-#include <dpsim-models/EMT/EMT_Ph1_Resistor.h>
+#include <dpsim-models/EMT/EMT_Ph3_ControlledCurrentSource.h>
+#include <dpsim-models/EMT/EMT_Ph3_Resistor.h>
 
 namespace CPS {
-namespace Signal {
+namespace EMT {
+namespace Ph3 {
 
-class DecouplingIdealTransformer_EMT_Ph1
+class DecouplingIdealTransformer
     : public CompositePowerComp<Real>,
-      public SharedFactory<DecouplingIdealTransformer_EMT_Ph1> {
+      public SharedFactory<DecouplingIdealTransformer> {
 protected:
   Real mDelay;
-  Real mInternalSeriesResistance = 1e-6;
-  Real mInternalParallelResistance = 1e6;
+  Matrix mInternalSeriesResistance =
+      CPS::Math::singlePhaseParameterToThreePhase(1e-6);
+  Matrix mInternalParallelResistance =
+      CPS::Math::singlePhaseParameterToThreePhase(1e6);
 
-  std::shared_ptr<EMT::Ph1::Resistor> mRes1, mRes2;
-  std::shared_ptr<EMT::Ph1::CurrentSource> mCurrentSrc;
-  std::shared_ptr<EMT::Ph1::VoltageSource> mVoltageSrc;
-  Attribute<Complex>::Ptr mSrcCurrent, mSrcVoltage;
+  std::shared_ptr<EMT::Ph3::Resistor> mRes1, mRes2;
+  std::shared_ptr<EMT::Ph3::ControlledCurrentSource> mCurrentSrc;
+  std::shared_ptr<EMT::Ph3::ControlledVoltageSource> mVoltageSrc;
+  Attribute<Matrix>::Ptr mSrcCurrent, mSrcVoltage;
 
   // Ringbuffers for the values of previous timesteps
   // TODO make these matrix attributes
-  std::vector<Real> mCur1, mVol2;
+  Matrix mCur1, mVol2;
 
   // Copy of the most recent elements of the ring buffers
   // They are used to perform extrapolation
-  std::vector<Real> mCur1Extrap, mVol2Extrap;
-  Real mCurrent1Extrap0;
+  Matrix mCur1Extrap, mVol2Extrap;
+  MatrixComp mCurrent1Extrap0;
 
   UInt mBufIdx = 0;
   UInt mMacroBufIdx = 0;
@@ -42,33 +46,33 @@ protected:
   Real mAlpha;
   CouplingMethod mCouplingMethod;
   UInt mExtrapolationDegree = 0;
-  Eigen::MatrixXd mVoltageSrcIntfCurr;
+  Matrix mVoltageSrcIntfCurr;
 
   // Get an approximate value of the signal in between steps when the delay is not an integer multiple of the step size
-  Real interpolate(std::vector<Real> &data);
+  Matrix interpolate(Matrix &data);
 
   // Estimates the value of the input signal in the next step
-  Real extrapolate(std::vector<Real> &data);
+  Matrix extrapolate(Matrix &data);
 
 public:
-  typedef std::shared_ptr<DecouplingIdealTransformer_EMT_Ph1> Ptr;
+  typedef std::shared_ptr<DecouplingIdealTransformer> Ptr;
 
-  const Attribute<Real>::Ptr mSourceVoltageIntfVoltage;
-  const Attribute<Real>::Ptr mSourceVoltageIntfCurrent;
-  const Attribute<Real>::Ptr mSrcCurrentRef;
-  const Attribute<Real>::Ptr mSrcVoltageRef;
+  const Attribute<Matrix>::Ptr mSourceVoltageIntfVoltage;
+  const Attribute<Matrix>::Ptr mSourceVoltageIntfCurrent;
+  const Attribute<Matrix>::Ptr mSrcCurrentRef;
+  const Attribute<Matrix>::Ptr mSrcVoltageRef;
 
   ///FIXME: workaround for dependency analysis as long as the states aren't attributes
   const Attribute<Matrix>::Ptr mStates;
 
-  DecouplingIdealTransformer_EMT_Ph1(
-      String uid, String name, Logger::Level logLevel = Logger::Level::info);
-  DecouplingIdealTransformer_EMT_Ph1(
-      String name, Logger::Level logLevel = Logger::Level::info)
-      : DecouplingIdealTransformer_EMT_Ph1(name, name, logLevel) {}
+  DecouplingIdealTransformer(String uid, String name,
+                             Logger::Level logLevel = Logger::Level::info);
+  DecouplingIdealTransformer(String name,
+                             Logger::Level logLevel = Logger::Level::info)
+      : DecouplingIdealTransformer(name, name, logLevel) {}
 
   void setParameters(Real delay, Matrix voltageSrcIntfCurr,
-                     Real current1Extrap0,
+                     Matrix current1Extrap0,
                      CouplingMethod method = CouplingMethod::DELAY);
   void step(Real time, Int timeStepCount);
   void postStep();
@@ -95,5 +99,6 @@ public:
   void mnaCompUpdateVoltage(const Matrix &leftVector) override;
   void mnaCompUpdateCurrent(const Matrix &leftVector) override;
 };
-} // namespace Signal
+} // namespace Ph3
+} // namespace EMT
 } // namespace CPS
