+++ dpsim-models/src/EMT/EMT_Ph1_DecouplingLine.cpp
@@ -6,26 +6,25 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  *********************************************************************************/
 
-#include "dpsim-models/Definitions.h"
-#include <dpsim-models/Signal/DecouplingLine.h>
+#include <dpsim-models/EMT/EMT_Ph1_DecouplingLine.h>
 
 using namespace CPS;
-using namespace CPS::DP::Ph1;
-using namespace CPS::Signal;
+using namespace CPS::EMT::Ph1;
 
-DecouplingLine::DecouplingLine(String uid, String name, Logger::Level logLevel)
-    : CompositePowerComp<Complex>(uid, name, true, true, logLevel),
+EMT::Ph1::DecouplingLine::DecouplingLine(String uid, String name,
+                                         Logger::Level logLevel)
+    : CompositePowerComp<Real>(uid, name, true, true, logLevel),
       mStates(mAttributes->create<Matrix>("states")),
-      mSrcCur1Ref(mAttributes->create<Complex>("i_src1")),
-      mSrcCur2Ref(mAttributes->create<Complex>("i_src2")) {
+      mSrcCur1Ref(mAttributes->create<Real>("i_src1")),
+      mSrcCur2Ref(mAttributes->create<Real>("i_src2")) {
 
   setTerminalNumber(2);
-  **mIntfVoltage = MatrixComp::Zero(1, 1);
-  **mIntfCurrent = MatrixComp::Zero(1, 1);
+  **mIntfVoltage = Matrix::Zero(1, 1);
+  **mIntfCurrent = Matrix::Zero(1, 1);
 }
 
-void DecouplingLine::setParameters(Real resistance, Real inductance,
-                                   Real capacitance) {
+void EMT::Ph1::DecouplingLine::setParameters(Real resistance, Real inductance,
+                                             Real capacitance) {
 
   mResistance = resistance;
   mInductance = inductance;
@@ -39,32 +38,32 @@ void DecouplingLine::setParameters(Real resistance, Real inductance,
   mParametersSet = true;
 }
 
-void DecouplingLine::createSubComponents() {
+void EMT::Ph1::DecouplingLine::createSubComponents() {
   if (mSubCompCreated)
     return;
   mSubCompCreated = true;
 
   mRes1 = Resistor::make(**mName + "_r1", mLogLevel);
   mRes1->setParameters(mSurgeImpedance + mResistance / 4);
-  mRes1->connect({mTerminals[0]->node(), SimNode<Complex>::GND});
+  mRes1->connect({mTerminals[0]->node(), EMT::SimNode::GND});
   addMNASubComponent(mRes1, MNA_SUBCOMP_TASK_ORDER::NO_TASK,
                      MNA_SUBCOMP_TASK_ORDER::TASK_BEFORE_PARENT, false);
 
   mRes2 = Resistor::make(**mName + "_r2", mLogLevel);
   mRes2->setParameters(mSurgeImpedance + mResistance / 4);
-  mRes2->connect({mTerminals[1]->node(), SimNode<Complex>::GND});
+  mRes2->connect({mTerminals[1]->node(), EMT::SimNode::GND});
   addMNASubComponent(mRes2, MNA_SUBCOMP_TASK_ORDER::NO_TASK,
                      MNA_SUBCOMP_TASK_ORDER::TASK_BEFORE_PARENT, false);
 
   mSrc1 = CurrentSource::make(**mName + "_i1", mLogLevel);
   mSrc1->setParameters(0);
-  mSrc1->connect({mTerminals[0]->node(), SimNode<Complex>::GND});
+  mSrc1->connect({mTerminals[0]->node(), EMT::SimNode::GND});
   addMNASubComponent(mSrc1, MNA_SUBCOMP_TASK_ORDER::NO_TASK,
                      MNA_SUBCOMP_TASK_ORDER::TASK_BEFORE_PARENT, true);
 
   mSrc2 = CurrentSource::make(**mName + "_i2", mLogLevel);
   mSrc2->setParameters(0);
-  mSrc2->connect({mTerminals[1]->node(), SimNode<Complex>::GND});
+  mSrc2->connect({mTerminals[1]->node(), EMT::SimNode::GND});
   addMNASubComponent(mSrc2, MNA_SUBCOMP_TASK_ORDER::NO_TASK,
                      MNA_SUBCOMP_TASK_ORDER::TASK_BEFORE_PARENT, true);
 
@@ -72,12 +71,14 @@ void DecouplingLine::createSubComponents() {
   mSrcCur2 = mSrc2->mCurrentRef;
 }
 
-void DecouplingLine::initializeParentFromNodesAndTerminals(Real frequency) {
-  (**mIntfVoltage)(0, 0) = initialSingleVoltage(1) - initialSingleVoltage(0);
+void EMT::Ph1::DecouplingLine::initializeParentFromNodesAndTerminals(
+    Real frequency) {
+  (**mIntfVoltage)(0, 0) =
+      (initialSingleVoltage(1) - initialSingleVoltage(0)).real();
 }
 
-void DecouplingLine::mnaParentInitialize(Real omega, Real timeStep,
-                                         Attribute<Matrix>::Ptr leftVector) {
+void EMT::Ph1::DecouplingLine::mnaParentInitialize(
+    Real omega, Real timeStep, Attribute<Matrix>::Ptr leftVector) {
   if (mDelay < timeStep)
     throw SystemError("Timestep too large for decoupling");
 
@@ -85,9 +86,9 @@ void DecouplingLine::mnaParentInitialize(Real omega, Real timeStep,
   mAlpha = 1 - (mBufSize - mDelay / timeStep);
   SPDLOG_LOGGER_INFO(mSLog, "bufsize {} alpha {}", mBufSize, mAlpha);
 
+  // Initialization based on static PI-line model
   Complex volt1 = initialSingleVoltage(0);
   Complex volt2 = initialSingleVoltage(1);
-  // TODO different initialization for lumped resistance?
   Complex initAdmittance = 1. / Complex(mResistance, omega * mInductance) +
                            Complex(0, omega * mCapacitance / 2);
   Complex cur1 = volt1 * initAdmittance -
@@ -97,36 +98,36 @@ void DecouplingLine::mnaParentInitialize(Real omega, Real timeStep,
   SPDLOG_LOGGER_INFO(mSLog, "initial voltages: v_k {} v_m {}", volt1, volt2);
   SPDLOG_LOGGER_INFO(mSLog, "initial currents: i_km {} i_mk {}", cur1, cur2);
 
-  (**mIntfCurrent)(0, 0) = cur1;
+  (**mIntfCurrent)(0, 0) = cur1.real();
 
   // Resize ring buffers and initialize
-  mVolt1.resize(mBufSize, volt1);
-  mVolt2.resize(mBufSize, volt2);
-  mCur1.resize(mBufSize, cur1);
-  mCur2.resize(mBufSize, cur2);
+  mVolt1.resize(mBufSize, volt1.real());
+  mVolt2.resize(mBufSize, volt2.real());
+  mCur1.resize(mBufSize, cur1.real());
+  mCur2.resize(mBufSize, cur2.real());
 }
 
-Complex DecouplingLine::interpolate(std::vector<Complex> &data) {
+Real EMT::Ph1::DecouplingLine::interpolate(std::vector<Real> &data) {
   // linear interpolation of the nearest values
-  Complex c1 = data[mBufIdx];
-  Complex c2 = mBufIdx == mBufSize - 1 ? data[0] : data[mBufIdx + 1];
+  Real c1 = data[mBufIdx];
+  Real c2 = mBufIdx == mBufSize - 1 ? data[0] : data[mBufIdx + 1];
   return mAlpha * c1 + (1 - mAlpha) * c2;
 }
 
-void DecouplingLine::step(Real time, Int timeStepCount) {
-  Complex volt1 = interpolate(mVolt1);
-  Complex volt2 = interpolate(mVolt2);
-  Complex cur1 = interpolate(mCur1);
-  Complex cur2 = interpolate(mCur2);
+void EMT::Ph1::DecouplingLine::step(Real time, Int timeStepCount) {
+  Real volt1 = interpolate(mVolt1);
+  Real volt2 = interpolate(mVolt2);
+  Real cur1 = interpolate(mCur1);
+  Real cur2 = interpolate(mCur2);
+  Real denom =
+      (mSurgeImpedance + mResistance / 4) * (mSurgeImpedance + mResistance / 4);
 
   if (timeStepCount == 0) {
-    // bit of a hack for proper initialization
+    // initialization
     **mSrcCur1Ref = cur1 - volt1 / (mSurgeImpedance + mResistance / 4);
     **mSrcCur2Ref = cur2 - volt2 / (mSurgeImpedance + mResistance / 4);
   } else {
     // Update currents
-    Real denom = (mSurgeImpedance + mResistance / 4) *
-                 (mSurgeImpedance + mResistance / 4);
     **mSrcCur1Ref = -mSurgeImpedance / denom *
                         (volt2 + (mSurgeImpedance - mResistance / 4) * cur2) -
                     mResistance / 4 / denom *
@@ -135,51 +136,47 @@ void DecouplingLine::step(Real time, Int timeStepCount) {
                         (volt1 + (mSurgeImpedance - mResistance / 4) * cur1) -
                     mResistance / 4 / denom *
                         (volt2 + (mSurgeImpedance - mResistance / 4) * cur2);
-    **mSrcCur1Ref = **mSrcCur1Ref * Complex(cos(-2. * PI * 50 * mDelay),
-                                            sin(-2. * PI * 50 * mDelay));
-    **mSrcCur2Ref = **mSrcCur2Ref * Complex(cos(-2. * PI * 50 * mDelay),
-                                            sin(-2. * PI * 50 * mDelay));
   }
   mSrcCur1->set(**mSrcCur1Ref);
   mSrcCur2->set(**mSrcCur2Ref);
 }
 
-void DecouplingLine::postStep() {
+void EMT::Ph1::DecouplingLine::postStep() {
   // Update ringbuffers with new values
   mVolt1[mBufIdx] = -mRes1->intfVoltage()(0, 0);
   mVolt2[mBufIdx] = -mRes2->intfVoltage()(0, 0);
-  mCur1[mBufIdx] = -mRes1->intfCurrent()(0, 0) + mSrcCur1->get();
-  mCur2[mBufIdx] = -mRes2->intfCurrent()(0, 0) + mSrcCur2->get();
+  mCur1[mBufIdx] = -mRes1->intfCurrent()(0, 0) + mSrcCur1->get().real();
+  mCur2[mBufIdx] = -mRes2->intfCurrent()(0, 0) + mSrcCur2->get().real();
 
   mBufIdx++;
   if (mBufIdx == mBufSize)
     mBufIdx = 0;
 }
 
-void DecouplingLine::mnaParentPreStep(Real time, Int timeStepCount) {
+void EMT::Ph1::DecouplingLine::mnaParentPreStep(Real time, Int timeStepCount) {
   step(time, timeStepCount);
   mSrc1->mnaPreStep(time, timeStepCount);
   mSrc2->mnaPreStep(time, timeStepCount);
   mnaCompApplyRightSideVectorStamp(**mRightVector);
 }
 
-void DecouplingLine::mnaParentPostStep(Real time, Int timeStepCount,
-                                       Attribute<Matrix>::Ptr &leftVector) {
+void EMT::Ph1::DecouplingLine::mnaParentPostStep(
+    Real time, Int timeStepCount, Attribute<Matrix>::Ptr &leftVector) {
   mnaCompUpdateVoltage(**leftVector);
   mnaCompUpdateCurrent(**leftVector);
   postStep();
 }
 
-void DecouplingLine::mnaCompUpdateVoltage(const Matrix &leftVector) {
+void EMT::Ph1::DecouplingLine::mnaCompUpdateVoltage(const Matrix &leftVector) {
   (**mIntfVoltage)(0, 0) =
       -mRes2->intfVoltage()(0, 0) + mRes1->intfVoltage()(0, 0);
 }
 
-void DecouplingLine::mnaCompUpdateCurrent(const Matrix &leftVector) {
-  (**mIntfCurrent)(0, 0) = -mRes1->intfCurrent()(0, 0) + mSrcCur1->get();
+void EMT::Ph1::DecouplingLine::mnaCompUpdateCurrent(const Matrix &leftVector) {
+  (**mIntfCurrent)(0, 0) = -mRes1->intfCurrent()(0, 0) + mSrcCur1->get().real();
 }
 
-void DecouplingLine::mnaParentAddPreStepDependencies(
+void EMT::Ph1::DecouplingLine::mnaParentAddPreStepDependencies(
     AttributeBase::List &prevStepDependencies,
     AttributeBase::List &attributeDependencies,
     AttributeBase::List &modifiedAttributes) {
@@ -187,7 +184,7 @@ void DecouplingLine::mnaParentAddPreStepDependencies(
   modifiedAttributes.push_back(mRightVector);
 }
 
-void DecouplingLine::mnaParentAddPostStepDependencies(
+void EMT::Ph1::DecouplingLine::mnaParentAddPostStepDependencies(
     AttributeBase::List &prevStepDependencies,
     AttributeBase::List &attributeDependencies,
     AttributeBase::List &modifiedAttributes,
