+++ dpsim-models/src/SP/SP_Ph1_DecouplingIdealTransformer.cpp
@@ -2,33 +2,31 @@
 // SPDX-License-Identifier: MPL-2.0
 
 #include "dpsim-models/Definitions.h"
+#include "dpsim-models/SimNode.h"
 #include "dpsim-models/TopologicalNode.h"
-#include <dpsim-models/Signal/DecouplingIdealTransformer_EMT_Ph3.h>
+#include <cstdlib>
+#include <dpsim-models/SP/SP_Ph1_DecouplingIdealTransformer.h>
 
 using namespace CPS;
-using namespace CPS::EMT::Ph3;
-using namespace CPS::Signal;
+using namespace CPS::SP::Ph1;
 
-DecouplingIdealTransformer_EMT_Ph3::DecouplingIdealTransformer_EMT_Ph3(
+SP::Ph1::DecouplingIdealTransformer::DecouplingIdealTransformer(
     String uid, String name, Logger::Level logLevel)
-    : CompositePowerComp<Real>(uid, name, true, true, logLevel),
+    : CompositePowerComp<Complex>(uid, name, true, true, logLevel),
       mStates(mAttributes->create<Matrix>("states")),
-      mSourceVoltageIntfVoltage(
-          mAttributes->create<Matrix>("v_src_intf", Matrix::Zero(3, 1))),
-      mSourceVoltageIntfCurrent(
-          mAttributes->create<Matrix>("i_src_intf", Matrix::Zero(3, 1))),
-      mSrcVoltageRef(mAttributes->create<Matrix>("v_ref", Matrix::Zero(3, 1))),
-      mSrcCurrentRef(mAttributes->create<Matrix>("i_ref", Matrix::Zero(3, 1))) {
+      mSourceVoltageIntfVoltage(mAttributes->create<Complex>("v_src_intf")),
+      mSourceVoltageIntfCurrent(mAttributes->create<Complex>("i_src_intf")),
+      mSrcVoltageRef(mAttributes->create<Complex>("v_ref")),
+      mSrcCurrentRef(mAttributes->create<Complex>("i_ref")) {
 
-  mPhaseType = PhaseType::ABC;
   setTerminalNumber(2);
   setVirtualNodeNumber(1);
-  **mIntfVoltage = Matrix::Zero(3, 1);
-  **mIntfCurrent = Matrix::Zero(3, 1);
+  **mIntfVoltage = MatrixComp::Zero(1, 1);
+  **mIntfCurrent = MatrixComp::Zero(1, 1);
 }
 
-void DecouplingIdealTransformer_EMT_Ph3::setParameters(
-    Real delay, Matrix voltageSrcIntfCurr, Matrix current1Extrap0,
+void SP::Ph1::DecouplingIdealTransformer::setParameters(
+    Real delay, Matrix voltageSrcIntfCurr, Complex current1Extrap0,
     CouplingMethod method) {
 
   mDelay = delay;
@@ -43,7 +41,7 @@ void DecouplingIdealTransformer_EMT_Ph3::setParameters(
   mParametersSet = true;
 }
 
-void DecouplingIdealTransformer_EMT_Ph3::createSubComponents() {
+void SP::Ph1::DecouplingIdealTransformer::createSubComponents() {
   if (mSubCompCreated)
     return;
   mSubCompCreated = true;
@@ -56,19 +54,19 @@ void DecouplingIdealTransformer_EMT_Ph3::createSubComponents() {
 
   mRes2 = Resistor::make(**mName + "_r2", mLogLevel);
   mRes2->setParameters(mInternalParallelResistance);
-  mRes2->connect({mTerminals[1]->node(), EMT::SimNode::GND});
+  mRes2->connect({mTerminals[1]->node(), CPS::SimNode<Complex>::GND});
   addMNASubComponent(mRes2, MNA_SUBCOMP_TASK_ORDER::NO_TASK,
                      MNA_SUBCOMP_TASK_ORDER::TASK_BEFORE_PARENT, false);
 
   mVoltageSrc = ControlledVoltageSource::make(**mName + "_v", mLogLevel);
-  mVoltageSrc->setParameters(Matrix::Zero(3, 1));
-  mVoltageSrc->connect({mVirtualNodes[0], EMT::SimNode::GND});
+  mVoltageSrc->setParameters(Complex(0, 0));
+  mVoltageSrc->connect({CPS::SimNode<Complex>::GND, mVirtualNodes[0]});
   addMNASubComponent(mVoltageSrc, MNA_SUBCOMP_TASK_ORDER::NO_TASK,
                      MNA_SUBCOMP_TASK_ORDER::TASK_BEFORE_PARENT, true);
 
   mCurrentSrc = ControlledCurrentSource::make(**mName + "_i", mLogLevel);
-  mCurrentSrc->setParameters(Matrix::Zero(3, 1));
-  mCurrentSrc->connect({mTerminals[1]->node(), EMT::SimNode::GND});
+  mCurrentSrc->setParameters(Complex(0, 0));
+  mCurrentSrc->connect({CPS::SimNode<Complex>::GND, mTerminals[1]->node()});
   addMNASubComponent(mCurrentSrc, MNA_SUBCOMP_TASK_ORDER::NO_TASK,
                      MNA_SUBCOMP_TASK_ORDER::TASK_BEFORE_PARENT, true);
 
@@ -76,25 +74,26 @@ void DecouplingIdealTransformer_EMT_Ph3::createSubComponents() {
   mSrcCurrent = mCurrentSrc->mCurrentRef;
 }
 
-void DecouplingIdealTransformer_EMT_Ph3::initializeParentFromNodesAndTerminals(
+void SP::Ph1::DecouplingIdealTransformer::initializeParentFromNodesAndTerminals(
     Real frequency) {
 
-  MatrixComp cur1 = mVoltageSrcIntfCurr.cast<Complex>();
-  MatrixComp volt2 = initialVoltage(1);
+  Complex cur1 = Complex(mVoltageSrcIntfCurr(0, 0), 0);
+  Complex volt2 = initialSingleVoltage(1) * RMS3PH_TO_PEAK1PH;
 
-  mVirtualNodes[0]->setInitialVoltage(initialVoltage(0) -
-                                      mInternalSeriesResistance * cur1);
+  mVirtualNodes[0]->setInitialVoltage(initialSingleVoltage(0) *
+                                          RMS3PH_TO_PEAK1PH -
+                                      cur1 * mInternalSeriesResistance);
 
   SPDLOG_LOGGER_INFO(mSLog, "initial current: i_1 {}", cur1);
   SPDLOG_LOGGER_INFO(mSLog, "initial voltage: v_2 {}", volt2);
 
-  **mSrcVoltageRef = volt2.real();
-  **mSrcCurrentRef = cur1.real();
+  **mSrcVoltageRef = volt2;
+  **mSrcCurrentRef = cur1;
   mVoltageSrc->setParameters(**mSrcVoltageRef);
   mCurrentSrc->setParameters(**mSrcCurrentRef);
 }
 
-void DecouplingIdealTransformer_EMT_Ph3::mnaParentInitialize(
+void SP::Ph1::DecouplingIdealTransformer::mnaParentInitialize(
     Real omega, Real timeStep, Attribute<Matrix>::Ptr leftVector) {
   if (mDelay <= 0) {
     mDelay = 0;
@@ -106,60 +105,62 @@ void DecouplingIdealTransformer_EMT_Ph3::mnaParentInitialize(
   }
   SPDLOG_LOGGER_INFO(mSLog, "bufsize {} alpha {}", mBufSize, mAlpha);
 
-  Matrix cur1 = **mSrcCurrentRef;
-  Matrix volt2 = **mSrcVoltageRef;
+  Complex cur1 = **mSrcCurrentRef;
+  Complex volt2 = **mSrcVoltageRef;
+  Matrix mSourceCurrentIntfVoltage(1, 1);
+  mSourceCurrentIntfVoltage(0, 0) = std::abs(volt2);
+  mCurrentSrc->setIntfVoltage(mSourceCurrentIntfVoltage);
 
-  mCurrentSrc->setIntfVoltage(volt2);
-
-  mVoltageSrc->setIntfVoltage(volt2);
+  mVoltageSrc->setIntfVoltage(mSourceCurrentIntfVoltage);
   mVoltageSrc->setIntfCurrent(mVoltageSrcIntfCurr);
 
   **mSourceVoltageIntfVoltage = volt2;
-  **mSourceVoltageIntfCurrent = mVoltageSrc->intfCurrent();
+  **mSourceVoltageIntfCurrent = mVoltageSrc->intfCurrent()(0, 0);
 
-  **mIntfVoltage = volt2;
-  **mIntfCurrent = cur1;
+  (**mIntfVoltage)(0, 0) = volt2;
+  (**mIntfCurrent)(0, 0) = cur1;
 
   // Resize ring buffers and initialize
-  mCur1 = cur1.transpose().replicate(mBufSize, 1);
-  mVol2 = volt2.transpose().replicate(mBufSize, 1);
+  mCur1.resize(mBufSize, cur1);
+  mVol2.resize(mBufSize, volt2);
 
   SPDLOG_LOGGER_INFO(mSLog, "Verify initial current: i_1 {}",
                      mCurrentSrc->intfCurrent()(0, 0));
   SPDLOG_LOGGER_INFO(mSLog, "Verify initial voltage: v_2 {}",
                      mVoltageSrc->intfVoltage()(0, 0));
 
-  mCur1Extrap = Matrix(mExtrapolationDegree + 1, 3);
-  mCur1Extrap.row(0) = mCurrent1Extrap0.real().transpose();
+  mCur1Extrap = std::vector<Complex>(mExtrapolationDegree + 1, Complex(0, 0));
+  mCur1Extrap[0] = mCurrent1Extrap0;
   if (mExtrapolationDegree > 0) {
-    mCur1Extrap.row(1) = mVoltageSrcIntfCurr.transpose();
+    mCur1Extrap[1] = mVoltageSrcIntfCurr(0, 0);
   }
-  mVol2Extrap = volt2.transpose().replicate(mExtrapolationDegree + 1, 1);
+  mVol2Extrap = std::vector<Complex>(mExtrapolationDegree + 1, volt2);
 }
 
-Matrix DecouplingIdealTransformer_EMT_Ph3::interpolate(Matrix &data) {
-  Matrix c1 = data.row(mBufIdx);
-  Matrix c2 = mBufIdx == mBufSize - 1 ? data.row(0) : data.row(mBufIdx + 1);
-  return (mAlpha * c1 + (1 - mAlpha) * c2).transpose();
+Complex
+SP::Ph1::DecouplingIdealTransformer::interpolate(std::vector<Complex> &data) {
+  Complex c1 = data[mBufIdx];
+  Complex c2 = mBufIdx == mBufSize - 1 ? data[0] : data[mBufIdx + 1];
+  return mAlpha * c1 + (1 - mAlpha) * c2;
 }
 
-Matrix DecouplingIdealTransformer_EMT_Ph3::extrapolate(Matrix &data) {
+Complex
+SP::Ph1::DecouplingIdealTransformer::extrapolate(std::vector<Complex> &data) {
   if (mCouplingMethod == CouplingMethod::EXTRAPOLATION_LINEAR) {
-    Matrix c1 = data.row(mMacroBufIdx);
-    Matrix c2 = mMacroBufIdx == mExtrapolationDegree
-                    ? data.row(0)
-                    : data.row(mMacroBufIdx + 1);
+    Complex c1 = data[mMacroBufIdx];
+    Complex c2 =
+        mMacroBufIdx == mExtrapolationDegree ? data[0] : data[mMacroBufIdx + 1];
     Real delayFraction =
         (mDelay * (mBufIdx + 1)) / static_cast<float>(mBufSize);
     Real tEval = mDelay + delayFraction;
-    return (((c2 - c1) / mDelay) * tEval + c1).transpose();
+    return ((c2 - c1) / mDelay) * tEval + c1;
   } else {
-    return (data.row(mMacroBufIdx)).transpose();
+    return data[mMacroBufIdx];
   }
 }
 
-void DecouplingIdealTransformer_EMT_Ph3::step(Real time, Int timeStepCount) {
-  Matrix volt1, cur2;
+void SP::Ph1::DecouplingIdealTransformer::step(Real time, Int timeStepCount) {
+  Complex volt1, cur2;
   if (mCouplingMethod == CouplingMethod::DELAY) {
     volt1 = interpolate(mVol2);
     cur2 = interpolate(mCur1);
@@ -171,22 +172,22 @@ void DecouplingIdealTransformer_EMT_Ph3::step(Real time, Int timeStepCount) {
   // Update voltage and current
   **mSrcVoltageRef = volt1;
   **mSrcCurrentRef = cur2;
-  **mSourceVoltageIntfVoltage = mVoltageSrc->intfVoltage();
-  **mSourceVoltageIntfCurrent = mVoltageSrc->intfCurrent();
+  **mSourceVoltageIntfVoltage = mVoltageSrc->intfVoltage()(0, 0);
+  **mSourceVoltageIntfCurrent = mVoltageSrc->intfCurrent()(0, 0);
 
   mSrcVoltage->set(**mSrcVoltageRef);
   mSrcCurrent->set(**mSrcCurrentRef);
 }
 
-void DecouplingIdealTransformer_EMT_Ph3::postStep() {
+void SP::Ph1::DecouplingIdealTransformer::postStep() {
   // Update ringbuffers with new values
-  mCur1.row(mBufIdx) = -mVoltageSrc->intfCurrent().transpose();
-  mVol2.row(mBufIdx) = mCurrentSrc->intfVoltage().transpose();
+  mCur1[mBufIdx] = mVoltageSrc->intfCurrent()(0, 0);
+  mVol2[mBufIdx] = -mCurrentSrc->intfVoltage()(0, 0);
 
   mBufIdx++;
   if (mBufIdx == mBufSize) {
-    mCur1Extrap.row(mMacroBufIdx) = mCur1.row(mBufIdx - 1);
-    mVol2Extrap.row(mMacroBufIdx) = mVol2.row(mBufIdx - 1);
+    mCur1Extrap[mMacroBufIdx] = mCur1[mBufIdx - 1];
+    mVol2Extrap[mMacroBufIdx] = mVol2[mBufIdx - 1];
     mMacroBufIdx++;
     if (mMacroBufIdx == mExtrapolationDegree + 1) {
       mMacroBufIdx = 0;
@@ -195,32 +196,32 @@ void DecouplingIdealTransformer_EMT_Ph3::postStep() {
   }
 }
 
-void DecouplingIdealTransformer_EMT_Ph3::mnaParentPreStep(Real time,
-                                                          Int timeStepCount) {
+void SP::Ph1::DecouplingIdealTransformer::mnaParentPreStep(Real time,
+                                                           Int timeStepCount) {
   step(time, timeStepCount);
   mVoltageSrc->mnaPreStep(time, timeStepCount);
   mCurrentSrc->mnaPreStep(time, timeStepCount);
   mnaCompApplyRightSideVectorStamp(**mRightVector);
 }
 
-void DecouplingIdealTransformer_EMT_Ph3::mnaParentPostStep(
+void SP::Ph1::DecouplingIdealTransformer::mnaParentPostStep(
     Real time, Int timeStepCount, Attribute<Matrix>::Ptr &leftVector) {
   mnaCompUpdateVoltage(**leftVector);
   mnaCompUpdateCurrent(**leftVector);
   postStep();
 }
 
-void DecouplingIdealTransformer_EMT_Ph3::mnaCompUpdateVoltage(
+void SP::Ph1::DecouplingIdealTransformer::mnaCompUpdateVoltage(
     const Matrix &leftVector) {
-  **mIntfVoltage = mVoltageSrc->intfVoltage();
+  (**mIntfVoltage)(0, 0) = mVoltageSrc->intfVoltage()(0, 0);
 }
 
-void DecouplingIdealTransformer_EMT_Ph3::mnaCompUpdateCurrent(
+void SP::Ph1::DecouplingIdealTransformer::mnaCompUpdateCurrent(
     const Matrix &leftVector) {
-  **mIntfCurrent = mVoltageSrc->intfCurrent();
+  (**mIntfCurrent)(0, 0) = mVoltageSrc->intfCurrent()(0, 0);
 }
 
-void DecouplingIdealTransformer_EMT_Ph3::mnaParentAddPreStepDependencies(
+void SP::Ph1::DecouplingIdealTransformer::mnaParentAddPreStepDependencies(
     AttributeBase::List &prevStepDependencies,
     AttributeBase::List &attributeDependencies,
     AttributeBase::List &modifiedAttributes) {
@@ -228,7 +229,7 @@ void DecouplingIdealTransformer_EMT_Ph3::mnaParentAddPreStepDependencies(
   modifiedAttributes.push_back(mRightVector);
 }
 
-void DecouplingIdealTransformer_EMT_Ph3::mnaParentAddPostStepDependencies(
+void SP::Ph1::DecouplingIdealTransformer::mnaParentAddPostStepDependencies(
     AttributeBase::List &prevStepDependencies,
     AttributeBase::List &attributeDependencies,
     AttributeBase::List &modifiedAttributes,
