+++ dpsim/src/pybind/SignalComponents.cpp
@@ -151,86 +151,6 @@ void addSignalComponents(py::module_ mSignal) {
           },
           "parameters"_a);
 
-  py::class_<CPS::Signal::DecouplingLine,
-             std::shared_ptr<CPS::Signal::DecouplingLine>,
-             CPS::SimPowerComp<CPS::Complex>>(mSignal, "DecouplingLine",
-                                              py::multiple_inheritance())
-      .def(py::init<std::string>())
-      .def(py::init<std::string, CPS::Logger::Level>())
-      .def("set_parameters", &CPS::Signal::DecouplingLine::setParameters,
-           "resistance"_a, "inductance"_a, "capacitance"_a)
-      .def("connect", &CPS::Signal::DecouplingLine::connect);
-
-  py::class_<CPS::Signal::DecouplingLineEMT,
-             std::shared_ptr<CPS::Signal::DecouplingLineEMT>,
-             CPS::SimPowerComp<CPS::Real>>(mSignal, "DecouplingLineEMT",
-                                           py::multiple_inheritance())
-      .def(py::init<std::string>())
-      .def(py::init<std::string, CPS::Logger::Level>())
-      .def("set_parameters", &CPS::Signal::DecouplingLineEMT::setParameters,
-           "resistance"_a, "inductance"_a, "capacitance"_a)
-      .def("connect", &CPS::Signal::DecouplingLineEMT::connect);
-
-  py::class_<CPS::Signal::DecouplingLineEMT_Ph3,
-             std::shared_ptr<CPS::Signal::DecouplingLineEMT_Ph3>,
-             CPS::SimPowerComp<CPS::Real>>(mSignal, "DecouplingLineEMT_Ph3",
-                                           py::multiple_inheritance())
-      .def(py::init<std::string>())
-      .def(py::init<std::string, CPS::Logger::Level>())
-      .def("set_parameters", &CPS::Signal::DecouplingLineEMT_Ph3::setParameters,
-           "resistance"_a, "inductance"_a, "capacitance"_a)
-      .def("connect", &CPS::Signal::DecouplingLineEMT_Ph3::connect);
-
-  py::class_<CPS::Signal::DecouplingIdealTransformer_SP_Ph1,
-             std::shared_ptr<CPS::Signal::DecouplingIdealTransformer_SP_Ph1>,
-             CPS::SimPowerComp<CPS::Complex>>(
-      mSignal, "DecouplingIdealTransformer_SP_Ph1", py::multiple_inheritance())
-      .def(py::init<std::string>())
-      .def(py::init<std::string, CPS::Logger::Level>())
-      .def("set_parameters",
-           &CPS::Signal::DecouplingIdealTransformer_SP_Ph1::setParameters,
-           "delay"_a, "v_src_intf_cur"_a, "cur1_extrap_0"_a,
-           "coupling_method"_a)
-      .def("connect", &CPS::Signal::DecouplingIdealTransformer_SP_Ph1::connect);
-
-  py::class_<CPS::Signal::DecouplingIdealTransformer_EMT_Ph1,
-             std::shared_ptr<CPS::Signal::DecouplingIdealTransformer_EMT_Ph1>,
-             CPS::SimPowerComp<CPS::Real>>(
-      mSignal, "DecouplingIdealTransformer_EMT_Ph1", py::multiple_inheritance())
-      .def(py::init<std::string>())
-      .def(py::init<std::string, CPS::Logger::Level>())
-      .def("set_parameters",
-           &CPS::Signal::DecouplingIdealTransformer_EMT_Ph1::setParameters,
-           "delay"_a, "v_src_intf_cur"_a, "cur1_extrap_0"_a,
-           "coupling_method"_a)
-      .def("connect",
-           &CPS::Signal::DecouplingIdealTransformer_EMT_Ph1::connect);
-
-  py::class_<CPS::Signal::DecouplingIdealTransformer_EMT_Ph3,
-             std::shared_ptr<CPS::Signal::DecouplingIdealTransformer_EMT_Ph3>,
-             CPS::SimPowerComp<CPS::Real>>(
-      mSignal, "DecouplingIdealTransformer_EMT_Ph3", py::multiple_inheritance())
-      .def(py::init<std::string>())
-      .def(py::init<std::string, CPS::Logger::Level>())
-      .def("set_parameters",
-           &CPS::Signal::DecouplingIdealTransformer_EMT_Ph3::setParameters,
-           "delay"_a, "v_src_intf_cur"_a, "cur1_extrap_0"_a,
-           "coupling_method"_a)
-      .def("connect",
-           &CPS::Signal::DecouplingIdealTransformer_EMT_Ph3::connect);
-
-  py::class_<CPS::Signal::DecouplingIdealTransformer_DP_Ph1,
-             std::shared_ptr<CPS::Signal::DecouplingIdealTransformer_DP_Ph1>,
-             CPS::SimPowerComp<CPS::Complex>>(
-      mSignal, "DecouplingIdealTransformer_DP_Ph1", py::multiple_inheritance())
-      .def(py::init<std::string>())
-      .def(py::init<std::string, CPS::Logger::Level>())
-      .def("set_parameters",
-           &CPS::Signal::DecouplingIdealTransformer_DP_Ph1::setParameters,
-           "delay"_a, "v_src_intf_cur"_a, "cur1_extrap_0"_a,
-           "coupling_method"_a)
-      .def("connect", &CPS::Signal::DecouplingIdealTransformer_DP_Ph1::connect);
-
   py::class_<CPS::Base::PSSParameters,
              std::shared_ptr<CPS::Base::PSSParameters>>(mSignal,
                                                         "PSSParameters");
